Update to support better image processing
